requirements.txt
numpy
torch
torchvision
opencv-python
matplotlib
pillow
imageio
imageio-ffmpeg
decord
xformers
einops
easydict
tqdm
OpenEXR==3.3.1
natsort